Revert "demux/packet: fix demux_packet_shorten"
This reverts commit 95636c65e73c1d0d8cba43d8c230291d99962e88.

This change shouldn't be needed, and in fact it's wrong. The FFmpeg API
function could do anything it wants with the packet, including changing
the packet data pointer. Likewise, it's not guaranteed that the
referenced packet's fields mirror the current state of the mpv packet
struct (the AVPacket is only kept for the AVBuffer and the side data
stuff).