DOCS: mention that mpv doesn't build with MSVC
And troll Microsoft slightly while we're at it. But is it trolling if
it's the truth?

The level of C99 support in MSVC is probably a bit better than most
people think, but it's by far not adequate. We need a bit of either C11
or GNU extensions too, and rely on some MinGW helpers (that look like
they're provided by MS, except they're not).