vo: fix forgotten debug code
This was not intended to be committed in 0e3f8936062967a9db. It disables
the extra wakeup if working==true. I've convinced myself that the wakeup
was really needed at the time, so no idea how I didn't notice this until
someone pointed it out on the commit diff on github (lol).