wayland: remove bogus scale_change variable
This was originally added in f2afae55e95b4b1eec1aeb828ba6ff1f0695d993
for unclear reasons (way to go me). This concept is clearly incorrect.
It doesn't matter what state the window is in. As soon as mpv detects a
scale change, it needs to reset the buffer scale of the window. Just
remove all this junk and put wl_surface_set_buffer_scale in
set_surface_scaling like it should be. Related issue: #9426.