player/video: check for track and decoder existence
The track during lavfi-complex can actually be NULL which meant that
ca4192e2df7bcfdb9e18461f19e1bd2dd0ee3c7a regressed lavfi-complex by
causing mpv to crash during runtime changes of the filter. Additionally,
it's possible for the decoder wrapper to also be NULL. check_framedrop
within write_video checks this, but check_for_hwdec_fallback does not.
Perhaps, it's impossible for this to happen, but we might as well add
the check here to be on the safe side since mp_decoder_wrapper_control
is not designed to handle a NULL.