Revert "win32/pthread: don't convert time through unrelated timer"
This reverts commit 318b5471a18e464cfcd1f7222da7853b7056f9fc.

While it may work, changing these two functions in violation of their documented
behaviour for the sake of a shortcut is a hack that will spell disaster sooner or later.

This is a partial revert since the commit in question also contained a hidden
bugfix where it swapped the calculation order for time_rel.