command: notify-property command
Since VOCTRL is not suitable for frequent data query
(see 477a0f83184d18a5cd89261b5f81ee668c7f4441 for details),
it's not suitable to be used by the VO clipboard backend.
Instead, since the VO does the clipboard monitoring by itself,
it can notify the player when the clipboard is updated.

This adds an internal notify-property command so that VOs
can notify player when the clipboard is updated, so that clipboard
monitoring works.