m_option: fix parsing of OP_APPEND for string lists
While parsing through the separators, the OP_APPEND also needs to use 0
for the separator since it is supposed to be unescaped. This was missed
in the code reorganization of d2c409c56b49252c85eefd340863fd59a36808d6.
Keyvalue lists were unaffected because only OP_DELETE and OP_REMOVE use
this code.