hwdec_d3d11va: align image mapper dimensions to a multiple of 2
This is required for creating a copy texture. Note that alignment is
applied only when making a copy; if mapping the texture directly, the
native texture size is used as-is. The texture itself may be larger
depending on the coded dimensions, but we only copy the "interesting"
part of the image.

Fixes playback of videos with odd dimensions.