Revert "manpage: clarify that you can prefetch m3u8 urls"
This commit is based on incorrect information, see previous commit.
It also deletes the correct information the prefetch starts when the
current file is fully load.

The commit also misunderstood the comment about m3u8 urls. It refers to
the prefetching of file data inside the m3u8 stream manifest data, which
has nothing to do with mpv playlists when the m3u8 urls are opened as one
stream by ffmpeg.

The original comment was not very clear about this, so add more details
about this.

This reverts commit d4bf179e78016f6145a7fa4a202c7f25887994e9.